extract reducer fn
